fix count

+++ client.go
@@ -957,7 +957,7 @@ func (mc *ModbusClient) ReadFileLines(recNumber uint16, quantity uint16) (values
 	log.Printf("startingRecord: %v", startingRecord)
 	log.Printf("startingRecord: %v", uint16ToBytes(BIG_ENDIAN, startingRecord))
 	for i := 1; i <= int(quantity); i++ {
-		queryString = append(queryString, 6)
+		queryString = append(queryString, uint8(6))
 		queryString = append(queryString, uint16ToBytes(BIG_ENDIAN, 1)...)
 		queryString = append(queryString, uint16ToBytes(BIG_ENDIAN, startingRecord)...)
 		queryString = append(queryString, uint16ToBytes(BIG_ENDIAN, RequestPayloadlength)...)
